Move functions used to create the separate item pages from App.js
